+++ src/net/tcp_transport_test.go
@@ -1,22 +1,20 @@
 package net
 
 import (
-	"net"
 	"testing"
 
 	"github.com/mosaicnetworks/babble/src/common"
 )
 
 func TestTCPTransport_BadAddr(t *testing.T) {
-	_, err := NewTCPTransport("0.0.0.0:0", nil, 1, 0, 0, common.NewTestEntry(t))
+	_, err := NewTCPTransport("0.0.0.0:0", "", 1, 0, 0, common.NewTestEntry(t))
 	if err != errNotAdvertisable {
 		t.Fatalf("err: %v", err)
 	}
 }
 
 func TestTCPTransport_WithAdvertise(t *testing.T) {
-	addr := &net.TCPAddr{IP: []byte{127, 0, 0, 1}, Port: 12345}
-	trans, err := NewTCPTransport("0.0.0.0:0", addr, 1, 0, 0, common.NewTestEntry(t))
+	trans, err := NewTCPTransport("0.0.0.0:0", "127.0.0.1:12345", 1, 0, 0, common.NewTestEntry(t))
 	if err != nil {
 		t.Fatalf("err: %v", err)
 	}