+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 ## Unreleased
 
 SECURITY:
+
 FEATURES:
+
+- net: Add configuration to advertise a different address than "BindAddr" to 
+       the rest of the cluster.
+
 IMPROVEMENTS:
+
 BUG FIXES:
 
 ## v0.5.5 (September 13, 2019)